Get the structure of sizes down,
but get too confused and so take a break. I knew that I could fold
over the children, but got confused exactly how these should call
each other or which sizes to use where. I took a break to relax
and mull this over in the back of my mind.