Finish part 1
With a clearer picture of what I wanted the sizes function to do,
and a much needed break, I started over with a new function and
figured out where I should use the sizes. I think I tried this at
one point in the previous attempt, but one quick fix is passing
the sizes that we have found along so that we can use them for sibling
calculations. Another thing that was important was passing the
initial sizes into `fold`.